Introduce "WithContext" variants of `FromRaw` and `AsRaw`
Based on PR comments, the ability to create Rust wrapper types from raw
`libudev` pointer types needs to be preserved.  Since doing this
correctly requires knowing not only the pointer to some `libudev` struct
but also the `struct udev*` that was provided as the context pointer
when that struct was allocated, I had to introduce `FromRawWithContext`
and `AsRawWithContext` and implement them on `Enumerator`, `Device`, and
`Builder`.