Add SocketIter and remove socket.clone()
This is a breaking patch that removes clone support for the Socket.
Clone is removed becuase it gives the user the assumption that cloning
the socket would allow user to receive multiple socket events from a
single udev monitor (aka single producer, multi consumer). However, the
udev monitor is a single producer, single consumer, and so cloning the
socket would have unexpected behavior.

Now, in order to iterate over socket events, you have the SocketIter
type, which implements Iterator and is internally a clone of the underlying
udev monitor. This clone is better here, because it is part of the
private API, and does not require caller to clone the socket when
iterating.

Signed-off-by: thomas <thomas.chiantia@gmail.com>