Export SocketIter and remove SocketIter cloneing
This fixes the SocketIter type (again) by holding a reference to the
socket (instead of cloneing). Doing this removes unecessary calls to
`ref` and `unref` when cloning and droping the iterator.

Also - the SocketIter type is exported as MonitorSocketIter. This allows
creating custom Iter types which can be used to map off the event types.

Signed-off-by: thomas <thomas.chiantia@gmail.com>